Исправления работы Sprint_7_Filippov_I.D
