+++ src/main/java/qa_scooter_praktikum/courier_and_order_pojo/CourierLogin.java
@@ -1,23 +1,13 @@
-package qa_scooter_praktikum;
+package qa_scooter_praktikum.courier_and_order_pojo;
 
-public class CourierRegistration {
+public class CourierLogin {
 
     private String login;
     private String password;
-    private String firstName;
 
-    public CourierRegistration(String login, String password, String firstName) {
+    public CourierLogin(String login, String password) {
         this.login = login;
         this.password = password;
-        this.firstName = firstName;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
     }
 
     public String getLogin() {
