+++ src/main/java/qa_scooter_praktikum/courier_and_order_pojo/Order.java
@@ -1,4 +1,4 @@
-package qa_scooter_praktikum;
+package qa_scooter_praktikum.courier_and_order_pojo;
 
 import java.util.List;
